requirements.txt
# Python packages required
ohbot
fuzzywuzzy
keyboard
Flask
azure-cognitiveservices-speech
opencv-python